+++ spotipy/oauth2.py
@@ -5,6 +5,8 @@ __all__ = [
     "SpotifyClientCredentials",
     "SpotifyOAuth",
     "SpotifyOauthError",
+    "SpotifyStateError",
+    "SpotifyImplicitGrant",
 ]
 
 import base64
@@ -29,10 +31,24 @@ logger = logging.getLogger(__name__)
 
 
 class SpotifyOauthError(Exception):
+    """ Error during Auth Code or Implicit Grant flow """
     def __init__(self, message, error=None, error_description=None, *args, **kwargs):
         self.error = error
         self.error_description = error_description
+        self.__dict__.update(kwargs)
         super(SpotifyOauthError, self).__init__(message, *args, **kwargs)
+
+
+class SpotifyStateError(SpotifyOauthError):
+    """ The state sent and state recieved were different """
+    def __init__(self, local_state=None, remote_state=None, message=None,
+                 error=None, error_description=None, *args, **kwargs):
+        if not message:
+            message = ("Expected " + local_state + " but recieved "
+                       + remote_state)
+        super(SpotifyOauthError, self).__init__(message, error,
+                                                error_description, *args,
+                                                **kwargs)
 
 
 def _make_authorization_headers(client_id, client_secret):
@@ -93,6 +109,13 @@ class SpotifyAuthBase(object):
     @redirect_uri.setter
     def redirect_uri(self, val):
         self._redirect_uri = _ensure_value(val, "redirect_uri")
+
+    @staticmethod
+    def _get_user_input(prompt):
+        try:
+            return raw_input(prompt)
+        except NameError:
+            return input(prompt)
 
     def __del__(self):
         """Make sure the connection (pool) gets closed"""
@@ -322,24 +345,21 @@ class SpotifyOAuth(SpotifyAuthBase):
             Parameters:
                 - url - the response url
         """
-        _, code, _ = self.parse_oauth_response_url(url)
+        _, code = self.parse_auth_response_url(url)
         if code is None:
             return url
         else:
             return code
 
     @staticmethod
-    def parse_oauth_response_url(url):
+    def parse_auth_response_url(url):
         query_s = urlparse(url).query
         form = dict(parse_qsl(query_s))
-        return tuple(form.get(param) for param in ['state', 'code', 'error'])
-
-    @staticmethod
-    def _get_user_input(prompt):
-        try:
-            return raw_input(prompt)
-        except NameError:
-            return input(prompt)
+        if "error" in form:
+            raise SpotifyOauthError("Received error from auth server: "
+                                    "{}".format(form["error"]),
+                                    error=form["error"])
+        return tuple(form.get(param) for param in ["state", "code"])
 
     def _make_authorization_headers(self):
         return _make_authorization_headers(self.client_id, self.client_secret)
@@ -355,9 +375,9 @@ class SpotifyOAuth(SpotifyAuthBase):
     def _get_auth_response_interactive(self):
         self._open_auth_url()
         response = SpotifyOAuth._get_user_input("Enter the URL you were redirected to: ")
-        state, code, _ = SpotifyOAuth.parse_oauth_response_url(response)
+        state, code = SpotifyOAuth.parse_auth_response_url(response)
         if self.state is not None and self.state != state:
-            raise SpotifyOauthError("Received inconsistent state from OAuth server.")
+            raise SpotifyStateError(self.state, state)
         return code
 
     def _get_auth_response_local_server(self, redirect_port):
@@ -366,7 +386,7 @@ class SpotifyOAuth(SpotifyAuthBase):
         server.handle_request()
 
         if self.state is not None and server.state != self.state:
-            raise SpotifyOauthError("Received inconsistent state from OAuth server.")
+            raise SpotifyStateError(self.state, server.state)
 
         if server.auth_code is not None:
             return server.auth_code
@@ -520,12 +540,251 @@ class SpotifyOAuth(SpotifyAuthBase):
         return token_info
 
 
+class SpotifyImplicitGrant(SpotifyAuthBase):
+    """ Implements Implicit Grant Flow for client apps
+
+    This auth manager enables *user and non-user* endpoints with only
+    a client secret, redirect uri, and username. The user will need to
+    copy and paste a URI from the browser every hour.
+
+    Security Advisory
+    -----------------
+    The Implicit Grant Flow is part of the
+    [OAuth 2.0 standard](https://oauth.net/2/grant-types/implicit/).
+    It is intended for client-side (running in browser or a native app)
+    interactions where the client secret would have to be hard-coded and
+    exposed. OAuth no longer recommends its use because sensitive
+    info (the auth token) can be yanked from the browser address bar or
+    history, instead recommending the Auth Code flow with PKCE. However,
+    Spotify [does not support PKCE](https://community.spotify.com/t5/Spotify-for-Developers/Authentication-API-failing-in-production-right-now/m-p/4960693/highlight/true#M234), <!---# noqa: E501-->
+    so Implicit Grant is the only viable options for client-side Spotify
+    API requests.
+    """
+    OAUTH_AUTHORIZE_URL = "https://accounts.spotify.com/authorize"
+
+    def __init__(self,
+                 client_id=None,
+                 redirect_uri=None,
+                 state=None,
+                 scope=None,
+                 cache_path=None,
+                 username=None,
+                 show_dialog=False):
+        """ Creates Auth Manager using the Implicit Grant flow
+
+        **See help(SpotifyImplictGrant) for Security Advisory**
+
+        Parameters
+        ----------
+        * client_id: Must be supplied or set as environment variable
+        * redirect_uri: Must be supplied or set as environment variable
+        * state: May be supplied, no verification is performed
+        * scope: May be supplied, intuitively converted to proper format
+        * cache_path: May be supplied, will otherwise be generated
+        * username: Must be supplied or set as environment variable
+        * show_dialog: Interpreted as boolean
+        """
+        self.client_id = client_id
+        self.redirect_uri = redirect_uri
+        self.state = state
+        self.cache_path = cache_path
+        self.username = username or os.getenv(
+            CLIENT_CREDS_ENV_VARS["client_username"]
+        )
+        self.scope = self._normalize_scope(scope)
+        self.show_dialog = show_dialog
+        self._session = None  # As to not break inherited __del__
+
+    def get_cached_token(self):
+        """ Gets a cached auth token
+        """
+        token_info = None
+
+        if not self.cache_path and self.username:
+            self.cache_path = ".cache-" + str(self.username)
+        elif not self.cache_path and not self.username:
+            raise SpotifyOauthError(
+                "You must either set a cache_path or a username."
+            )
+
+        if self.cache_path:
+            try:
+                f = open(self.cache_path)
+                token_info_string = f.read()
+                f.close()
+                token_info = json.loads(token_info_string)
+
+                # if scopes don't match, then bail
+                if "scope" not in token_info or not self._is_scope_subset(
+                    self.scope, token_info["scope"]
+                ):
+                    return None
+
+                if self.is_token_expired(token_info):
+                    return None
+
+            except IOError:
+                pass
+        return token_info
+
+    def _save_token_info(self, token_info):
+        if not self.cache_path and self.username:
+            self.cache_path = ".cache-" + str(self.username)
+        if self.cache_path:
+            try:
+                f = open(self.cache_path, "w")
+                f.write(json.dumps(token_info))
+                f.close()
+            except IOError:
+                logger.warning('Couldn\'t write token to cache at: %s',
+                               self.cache_path)
+
+    def _is_scope_subset(self, needle_scope, haystack_scope):
+        needle_scope = set(needle_scope.split()) if needle_scope else set()
+        haystack_scope = (
+            set(haystack_scope.split()) if haystack_scope else set()
+        )
+        return needle_scope <= haystack_scope
+
+    def is_token_expired(self, token_info):
+        return is_token_expired(token_info)
+
+    def get_access_token(self,
+                         state=None,
+                         response=None,
+                         check_cache=True):
+        """ Gets Auth Token from cache (preferred) or user interaction
+
+        Parameters
+        ----------
+        * state: May be given, overrides (without changing) self.state
+        * response: URI with token, can break expiration checks
+        * check_cache: Interpreted as boolean
+        """
+        if check_cache:
+            token_info = self.get_cached_token()
+            if not (token_info is None or is_token_expired(token_info)):
+                return token_info["access_token"]
+
+        if response:
+            token_info = self.parse_response_token(response)
+        else:
+            token_info = self.get_auth_response(state)
+        token_info = self._add_custom_values_to_token_info(token_info)
+        self._save_token_info(token_info)
+
+        return token_info["access_token"]
+
+    def _normalize_scope(self, scope):
+        if scope:
+            scopes = sorted(scope.split())
+            return " ".join(scopes)
+        else:
+            return None
+
+    def get_authorize_url(self, state=None):
+        """ Gets the URL to use to authorize this app """
+        payload = {
+            "client_id": self.client_id,
+            "response_type": "token",
+            "redirect_uri": self.redirect_uri,
+        }
+        if self.scope:
+            payload["scope"] = self.scope
+        if state is None:
+            state = self.state
+        if state is not None:
+            payload["state"] = state
+        if self.show_dialog:
+            payload["show_dialog"] = True
+
+        urlparams = urllibparse.urlencode(payload)
+
+        return "%s?%s" % (self.OAUTH_AUTHORIZE_URL, urlparams)
+
+    def parse_response_token(self, url, state=None):
+        """ Parse the response code in the given response url """
+        remote_state, token, t_type, exp_in = self.parse_auth_response_url(url)
+        if state is None:
+            state = self.state
+        if state is not None and remote_state != state:
+            raise SpotifyStateError(state, remote_state)
+        return {"access_token": token, "token_type": t_type,
+                "expires_in": exp_in, "state": state}
+
+    @staticmethod
+    def parse_auth_response_url(url):
+        url_components = urlparse(url)
+        fragment_s = url_components.fragment
+        query_s = url_components.query
+        form = dict(i.split('=') for i
+                    in (fragment_s or query_s or url).split('&'))
+        if "error" in form:
+            raise SpotifyOauthError("Received error from auth server: "
+                                    "{}".format(form["error"]),
+                                    state=form["state"])
+        if "expires_in" in form:
+            form["expires_in"] = int(form["expires_in"])
+        return tuple(form.get(param) for param in ["state", "access_token",
+                                                   "token_type", "expires_in"])
+
+    def _open_auth_url(self, state=None):
+        auth_url = self.get_authorize_url(state)
+        try:
+            webbrowser.open(auth_url)
+            logger.info("Opened %s in your browser", auth_url)
+        except webbrowser.Error:
+            logger.error("Please navigate here: %s", auth_url)
+
+    def get_auth_response(self, state=None):
+        """ Gets a new auth **token** with user interaction """
+        logger.info('User authentication requires interaction with your '
+                    'web browser. Once you enter your credentials and '
+                    'give authorization, you will be redirected to '
+                    'a url.  Paste that url you were directed to to '
+                    'complete the authorization.')
+
+        redirect_info = urlparse(self.redirect_uri)
+        redirect_host, redirect_port = get_host_port(redirect_info.netloc)
+        # Implicit Grant tokens are returned in a hash fragment
+        # which is only available to the browser. Therefore, interactive
+        # URL retrival is required.
+        if (redirect_host in ("127.0.0.1", "localhost")
+                and redirect_info.scheme == "http" and redirect_port):
+            logger.warning('Using a local redirect URI with a '
+                           'port, likely expecting automatic '
+                           'retrieval. Due to technical limitations, '
+                           'the authentication token cannot be '
+                           'automatically retrieved and must be '
+                           'copied and pasted.')
+
+        self._open_auth_url(state)
+        logger.info('Paste that url you were directed to in order to '
+                    'complete the authorization')
+        response = SpotifyImplicitGrant._get_user_input("Enter the URL you "
+                                                        "were redirected to: ")
+        return self.parse_response_token(response, state)
+
+    def _add_custom_values_to_token_info(self, token_info):
+        """
+        Store some values that aren't directly provided by a Web API
+        response.
+        """
+        token_info["expires_at"] = int(time.time()) + token_info["expires_in"]
+        token_info["scope"] = self.scope
+        return token_info
+
+
 class RequestHandler(BaseHTTPRequestHandler):
     def do_GET(self):
-        state, auth_code, error = SpotifyOAuth.parse_oauth_response_url(self.path)
-        self.server.state = state
-        self.server.auth_code = auth_code
-        self.server.error = error
+        self.server.auth_code = self.server.error = None
+        try:
+            state, auth_code = SpotifyOAuth.parse_auth_response_url(self.path)
+            self.server.state = state
+            self.server.auth_code = auth_code
+        except SpotifyOauthError as err:
+            self.server.state = err.state
+            self.server.error = err.error
 
         self.send_response(200)
         self.send_header("Content-Type", "text/html")
@@ -560,5 +819,6 @@ def start_local_http_server(port, handler=RequestHandler):
     server = HTTPServer(("127.0.0.1", port), handler)
     server.allow_reuse_address = True
     server.auth_code = None
+    server.auth_token_form = None
     server.error = None
     return server