+++ tests/integration/test_user_endpoints.py
@@ -5,6 +5,7 @@ from spotipy import (
     prompt_for_user_token,
     Spotify,
     SpotifyException,
+    SpotifyImplicitGrant
 )
 import unittest
 import requests
@@ -395,3 +396,36 @@ class SpotipyPlayerApiTests(unittest.TestCase):
             after=res['cursors']['before'])
         self.assertLessEqual(len(res['items']), 50)
         self.assertGreater(res['items'][0]['played_at'], played_at)
+
+
+class SpotipyImplicitGrantTests(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        scope = (
+            'user-follow-read '
+            'user-follow-modify '
+        )
+        auth_manager = SpotifyImplicitGrant(scope=scope,
+                                            cache_path=".cache-implicittest")
+        cls.spotify = Spotify(auth_manager=auth_manager)
+
+    def test_user_follows_and_unfollows_artist(self):
+        # Initially follows 1 artist
+        current_user_followed_artists = self.spotify.current_user_followed_artists()[
+            'artists']['total']
+
+        # Follow 2 more artists
+        artists = ["6DPYiyq5kWVQS4RGwxzPC7", "0NbfKEOTQCcwd6o7wSDOHI"]
+        self.spotify.user_follow_artists(artists)
+        res = self.spotify.current_user_followed_artists()
+        self.assertEqual(res['artists']['total'], current_user_followed_artists + len(artists))
+
+        # Unfollow these 2 artists
+        self.spotify.user_unfollow_artists(artists)
+        res = self.spotify.current_user_followed_artists()
+        self.assertEqual(res['artists']['total'], current_user_followed_artists)
+
+    def test_current_user(self):
+        c_user = self.spotify.current_user()
+        user = self.spotify.user(c_user['id'])
+        self.assertEqual(c_user['display_name'], user['display_name'])